fix: break code output retention import cycle
