+++ api/server/services/Files/process.js
@@ -8,7 +8,6 @@ const {
   FileContext,
   FileSources,
   imageExtRegex,
-  RetentionMode,
   EModelEndpoint,
   EToolResources,
   mergeFileConfig,
@@ -20,7 +19,7 @@ const {
   getEndpointFileConfig,
   documentParserMimeTypes,
 } = require('librechat-data-provider');
-const { logger, runAsSystem, createTempChatExpirationDate } = require('@librechat/data-schemas');
+const { logger, runAsSystem } = require('@librechat/data-schemas');
 const {
   sanitizeFilename,
   parseText,
@@ -38,71 +37,11 @@ const { loadAuthValues } = require('~/server/services/Tools/credentials');
 const { getFileStrategy } = require('~/server/utils/getFileStrategy');
 const { checkCapability } = require('~/server/services/Config');
 const { LB_QueueAsyncCall } = require('~/server/utils/queue');
+const { getRetentionExpiry } = require('./retention');
 const { getStrategyFunctions } = require('./strategies');
 const { determineFileType } = require('~/server/utils');
 const { STTService } = require('./Audio/STTService');
 const db = require('~/models');
-
-const isTruthy = (value) => value === true || value === 'true';
-
-const getConversationExpirationDate = (convo) => {
-  if (convo?.expiredAt == null) {
-    return null;
-  }
-
-  const expiredAt = convo.expiredAt instanceof Date ? convo.expiredAt : new Date(convo.expiredAt);
-  return Number.isNaN(expiredAt.getTime()) ? null : expiredAt;
-};
-
-const isActiveExpirationDate = (expiredAt) => expiredAt > new Date();
-
-function createRetentionExpiry(req) {
-  try {
-    return { expiredAt: createTempChatExpirationDate(req.config?.interfaceConfig) };
-  } catch (err) {
-    logger.error('[getRetentionExpiry] Error creating file expiration date:', err);
-    return { expiredAt: null };
-  }
-}
-
-/**
- * Returns `{ expiredAt }` when the request indicates data retention applies, otherwise `{}`.
- * Spread into file data objects before calling createFile.
- * @param {ServerRequest} req
- * @returns {Promise<{ expiredAt?: Date | null }>}
- */
-async function getRetentionExpiry(req) {
-  if (req?.config?.interfaceConfig?.retentionMode === RetentionMode.ALL) {
-    return createRetentionExpiry(req);
-  }
-
-  const conversationId = req?.body?.conversationId;
-  if (conversationId && req?.user?.id) {
-    try {
-      const convo = await db.getConvo(req.user.id, conversationId);
-      if (convo) {
-        const expiredAt = getConversationExpirationDate(convo);
-        if (expiredAt == null) {
-          return {};
-        }
-
-        if (!isActiveExpirationDate(expiredAt)) {
-          return { expiredAt };
-        }
-
-        return createRetentionExpiry(req);
-      }
-    } catch (err) {
-      logger.error('[getRetentionExpiry] Error checking conversation retention:', err);
-    }
-  }
-
-  if (!isTruthy(req?.body?.isTemporary)) {
-    return {};
-  }
-
-  return createRetentionExpiry(req);
-}
 
 /**
  * Creates a modular file upload wrapper that ensures filename sanitization
